pi-provider: remove receipt verification; prevention-only via SPKI pinning
Per product direction, the pi provider ships PREVENTION only. Attested TLS
(SPKI) pinning is the security control; per-response receipt verification
was an audit trail and is removed.

Removed:
- classifyReceipt / verifyReceipt / isFullyVerified classification, the
  message_end + after_provider_response receipt handlers, and the footer
  verified*/routed/mismatch states;
- fetchReceipt / fetchSession / body-hash (canonicalRequestBytes) plumbing;
- autoFetchReceipt and requireAttestationMatch config + settings toggles;
- the attestation-store (ne) receipt-tracking fields (renamed from
  receipt-store.ts; it now only caches the validated report for the pin);
- recompletion tests for receipt signature/body-hash/footer semantics.

Kept (prevention):
- @phala/aci-verifier verifyReportBinding for attestation binding (the pin is
  established only from a report that passes), SPKI extraction, fail-closed
  TLS pinning (tls-pinning.ts), failOpenOnUnpinned opt-out, /attestation.

Docs (core + shell READMEs + clients/README + PLAN note) reframed: the
plugin exists to keep traffic readable only by the attested workload;
receipt audit is delegated to @phala/aci-verifier for anyone who wants it.

tsc clean, oxlint clean, 34/34 node tests.
