+++ clients/pi-provider/packages/pi-provider-aci/src/settings-ui.ts
@@ -68,7 +68,6 @@ export function modelRegistrationSummary(config: AciCloudConfig): string {
 }
 
 export function verifySummary(config: AciCloudConfig): string {
-  return `Verify: auto-receipt ${formatBoolean(config.verify.autoFetchReceipt)}, ` +
-    `require-attestation ${formatBoolean(config.verify.requireAttestationMatch)}, ` +
+  return `Pin: ${config.pinning.enabled ? "required" : "disabled"}, ` +
     `fail-open-unpinned ${formatBoolean(config.verify.failOpenOnUnpinned)}`;
 }
